check -read-data - new tests
split the tests into 2 files:
internal/repository/check_test.go: runs TestGapInBlobs()

internal/checker/checker_test.go:
contains tests TestCheckPackHashMismatch(), TestCheckPackDownloadError() and
TestCheckPackPartialDownloadError()
