+++ internal/checker/checker_test.go
@@ -1,7 +1,6 @@
 package checker_test
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"io"
@@ -14,7 +13,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/klauspost/compress/zstd"
 	"github.com/restic/restic/internal/archiver"
 	"github.com/restic/restic/internal/backend"
 	"github.com/restic/restic/internal/checker"
@@ -570,51 +568,6 @@ func TestCheckerBlobTypeConfusion(t *testing.T) {
 		t.Fatal("no error found, checker is broken")
 	}
 	test.Assert(t, delayRepo.Triggered, "delay repository did not trigger")
-}
-
-const maxStreamBufferSize = 1 << (2 + 10 + 10) // 4 * 1024 * 1024 = 4 MiB
-
-func testWrapCheckPack(ctx context.Context, t *testing.T, repo *repository.Repository,
-	packID restic.ID, blobs []restic.Blob, size int64,
-) error {
-	bufRd := bufio.NewReaderSize(nil, maxStreamBufferSize)
-	dec, err := zstd.NewReader(nil)
-	test.OK(t, err)
-	return repository.CheckPack(ctx, repo, packID, blobs, size, bufRd, dec)
-}
-
-func TestGapInBlobs(t *testing.T) {
-	repo, _, cleanup := repository.TestFromFixture(t, checkerTestData)
-	defer cleanup()
-
-	chkr := checker.New(repo, true)
-	_, errs := chkr.LoadIndex(context.TODO(), nil)
-	test.Assert(t, len(errs) == 0, "expected no errors loading index")
-
-	repoPacks, err := pack.Size(context.TODO(), repo, false)
-	test.OK(t, err)
-
-	packID := restic.TestParseID("19a731a515618ec8b75fc0ff3b887d8feb83aef1001c9899f6702761142ed068")
-	_, ok := repoPacks[packID]
-	test.Assert(t, ok, "expected pack 19a731a515618ec8b75fc0ff3b887d8feb83aef1001c9899f6702761142ed068")
-
-	blobs := []restic.Blob{}
-	pb := <-repo.ListPacksFromIndex(context.TODO(), restic.NewIDSet(packID))
-	blobs = append(blobs, pb.Blobs...)
-
-	// force fail
-	blobs = blobs[1:]
-	err = testWrapCheckPack(context.TODO(), t, repo, packID, blobs, repoPacks[packID])
-
-	var packErr *repository.ErrPackData
-	test.Assert(t, errors.As(err, &packErr), "expected ErrPackData, got: %T %v", err, err)
-	test.Equals(t, packID, packErr.PackID)
-
-	errText := err.Error()
-	test.Assert(t, strings.Contains(errText, "gaps") || strings.Contains(errText, "overlapping"),
-		"expected gap/overlap error in: %v", errText)
-	test.Assert(t, strings.Contains(errText, "pack header size does not match"),
-		"expected header size mismatch error in: %v", errText)
 }
 
 // these test were generated by claude.ai, sonnet 4.6
